requirements.txt
# Core (sync scripts, APIs)
requests>=2.28.0
google-api-core>=2.11.0
python-dotenv>=1.0.0

# Meltano (optional: pip install -r requirements.txt)
meltano>=3.0,<5
